Movement
Player and Camera movement mostly complete, I still need to add more movement like a jump and sprint but for now this is my foundation to work with.

Camera zoom is subject to change, I'm not a big fan but I don't want to spend forever on it right now, aside from that I very much enjoy it.